+++ lib/tests/test_revset.rs
@@ -27,6 +27,7 @@ use jj_lib::backend::Signature;
 use jj_lib::backend::Timestamp;
 use jj_lib::commit::Commit;
 use jj_lib::default_index::DefaultIndexStore;
+use jj_lib::fileset::FilesetAliasesMap;
 use jj_lib::fileset::FilesetExpression;
 use jj_lib::git;
 use jj_lib::graph::GraphEdge;
@@ -100,6 +101,7 @@ fn resolve_symbol(repo: &dyn Repo, symbol: &str) -> Result<Vec<CommitId>, Revset
         user_email: "",
         date_pattern_context: chrono::Local::now().into(),
         default_ignored_remote: Some(git::REMOTE_NAME_FOR_LOCAL_GIT_REPO),
+        fileset_aliases_map: &FilesetAliasesMap::new(),
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: None,
@@ -233,6 +235,7 @@ fn test_resolve_symbol_commit_id() {
         user_email: settings.user_email(),
         date_pattern_context: chrono::Utc::now().fixed_offset().into(),
         default_ignored_remote: Some(git::REMOTE_NAME_FOR_LOCAL_GIT_REPO),
+        fileset_aliases_map: &FilesetAliasesMap::new(),
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: None,
@@ -1180,6 +1183,7 @@ fn try_resolve_expression(
         user_email: settings.user_email(),
         date_pattern_context: chrono::Utc::now().fixed_offset().into(),
         default_ignored_remote: Some(git::REMOTE_NAME_FOR_LOCAL_GIT_REPO),
+        fileset_aliases_map: &FilesetAliasesMap::new(),
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: None,
@@ -1231,6 +1235,7 @@ fn resolve_commit_ids_in_workspace(
         user_email: settings.user_email(),
         date_pattern_context: chrono::Utc::now().fixed_offset().into(),
         default_ignored_remote: Some(git::REMOTE_NAME_FOR_LOCAL_GIT_REPO),
+        fileset_aliases_map: &FilesetAliasesMap::new(),
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: Some(workspace_ctx),